Remove use of Categories from Themes.
